Se crearon la libreria de las banderas con su funcion, se modifico el main agregando un arreglo de 4 variables para las banderas y algunas correciones del tipo de variable en los registros

+++ registros.c
@@ -1,10 +1,11 @@
 #include "registros.h"
+#include <stdint.h>
 
-void mostrar_registro(unsigned long *reg)
+void mostrar_registro(uint32_t *reg)
 {
 	int i;
 	for(i=0;i<=12;i++)
 	{
-		printf("R%d: %d\n", i, reg[i]);
+		printf("R%d: %d\n", i,reg[i]);
 	}
 }
